refactor(memory): extract ProceduralMemoryMixin into procedural_memory.py (#887)
